Removed warnings

+++ Sources/ThumbsViewController.m
@@ -303,7 +303,7 @@
 
 	NSInteger page = (showBookmarked ? [[bookmarked objectAtIndex:index] integerValue] : (index + 1));
 
-	[thumbCell showText:[NSString stringWithFormat:@"%i", page]]; // Page number place holder
+	[thumbCell showText:[NSString stringWithFormat:@"%i", (int)page]]; // Page number place holder
 
 	[thumbCell showBookmark:[document.bookmarks containsIndex:page]]; // Show bookmarked status
